+++ MyLayout/RLTest5ViewController.h
@@ -1,0 +1,13 @@
+//
+//  RLTest5ViewController.h
+//  MyLayout
+//
+//  Created by oybq on 16/12/19.
+//  Copyright (c) 2016年 YoungSoft. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface RLTest5ViewController : UIViewController
+
+@end